[SDSAET-38] Move AccessibilityError models to models-api module
