+++ api/models-api/NOTICE
@@ -1,0 +1,20 @@
+Cognifide Limited
+Copyright 2013, Cognifide Limited
+This project includes software developed by Cognifide.
+http://www.cognifide.com
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance
+with the License. You may obtain a copy of the License at:
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing,
+software distributed under the License is distributed on
+an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+KIND, either express or implied. See the License for the
+specific language governing permissions and limitations
+under the License.
+
+This project includes:
+  net.lightbody.bmp:browsermob-proxy:2.0.0 (core packages) under The Apache Software License, Version 2.0 from http://bmp.lightbody.net/
