+++ api/models-api/src/main/java/com/cognifide/aet/models/AccessibilityIssue.java
@@ -1,10 +1,10 @@
-/**
+/*
  * AET
  *
  * Copyright (C) 2013 Cognifide Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ *  in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -13,9 +13,8 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.job.common.collectors.accessibility;
+package com.cognifide.aet.models;
 
-import com.cognifide.aet.job.common.Excludable;
 import java.io.Serializable;
 import java.util.Objects;
