+++ core/jobs/src/main/java/com/cognifide/aet/job/common/collectors/accessibility/AccessibilityIssueMarkupFinder.java
@@ -15,6 +15,7 @@
  */
 package com.cognifide.aet.job.common.collectors.accessibility;
 
+import com.cognifide.aet.models.AccessibilityIssue;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
