+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/accessibility/AccessibilityComparator.java
@@ -21,11 +21,11 @@ import com.cognifide.aet.job.api.comparator.ComparatorProperties;
 import com.cognifide.aet.job.api.datafilter.DataFilterJob;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
-import com.cognifide.aet.job.common.Excludable;
-import com.cognifide.aet.job.common.collectors.accessibility.AccessibilityIssue;
-import com.cognifide.aet.job.common.comparators.accessibility.report.AccessibilityReport;
-import com.cognifide.aet.job.common.comparators.accessibility.report.AccessibilityReportConfiguration;
 import com.cognifide.aet.job.common.comparators.accessibility.report.AccessibilityReportGenerator;
+import com.cognifide.aet.models.AccessibilityIssue;
+import com.cognifide.aet.models.AccessibilityReport;
+import com.cognifide.aet.models.AccessibilityReportConfiguration;
+import com.cognifide.aet.models.Excludable;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Lists;
