+++ rest-endpoint/src/main/java/com/cognifide/aet/models/AccessibilityErrorWrapper.java
@@ -15,9 +15,6 @@
  */
 package com.cognifide.aet.models;
 
-import com.cognifide.aet.job.common.collectors.accessibility.AccessibilityIssue;
-import com.cognifide.aet.job.common.comparators.accessibility.report.AccessibilityReport;
-import com.cognifide.aet.job.common.comparators.accessibility.report.AccessibilityReportConfiguration;
 import com.google.gson.reflect.TypeToken;
 import java.lang.reflect.Type;
 import java.util.List;
